pyproject.toml
[project]
name = "pi-tts-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coqui-tts>=0.26.2",
    "py-cord>=2.6.1",
    "torch>=2.7.1",
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.0.0",
    "dotenv>=0.9.9",
    "pyyaml>=6.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]

# [tool.mypy]
# python_version = "3.13"
# warn_return_any = true
# warn_unused_configs = true
# disallow_untyped_defs = true
# disallow_incomplete_defs = true
# check_untyped_defs = true
# disallow_untyped_decorators = true
# no_implicit_optional = true
# warn_redundant_casts = true
# warn_unused_ignores = true
# warn_no_return = true
# warn_unreachable = true
# strict_equality = true

# [[tool.mypy.overrides]]
# module = [
#     "discord.*",
#     "TTS.*",
#     "yaml.*",
# ]
# ignore_missing_imports = true
